Provisional template brain object for BANC

This is calibrated in nm. See data-raw/BANC.R for details of how it was generated. It is still provisional pending synpase cloud-based registrations.

Usage

BANC

Format

An object of class templatebrain of dimension 262144 x 294912 x 7010.

Examples

BANC
#> === Template Brain === 
#> Name: BANC 
#> Short Name: BANC 
#> Type: single brain 
#> Sex:  F 
#> Dimensions:262144 x 294912 x 7010 voxels
#> Voxel size:
#>   x = 4 nm
#>   y = 4 NA
#>   z = 45 NA
#> Bounding box (nm):
#>   x = , y = , z = ,
#>   x = , y = , z = .
#> Description: The BANC (said 'the bank') is the Brain And Nerve Cord, a GridTape transmission
#> electron microscopy dataset of a female adult Drosophila melanogaster's entire
#> central nervous system. This dataset was generated by Jasper Phelps and 
#> Minsu Kim in Wei-Chung Lee's lab at Harvard Medical School.
#> 3D segmentation to extract the shapes of neurons and other objects in the
#> dataset was performed by Zetta AI. The community proofreading and annotation
#> effort is coordinated by FlyWire.
#> See https://github.com/jasper-tms/the-BANC-fly-connectome/wiki for details. 
#> DOI:
# Depends on nat
library(nat)
boundingbox(BANC)
#>         [,1]    [,2]   [,3]
#> [1,]       0       0      0
#> [2,] 1048572 1179644 315405
#> attr(,"class")
#> [1] "boundingbox"