Low level access to FANC's CAVE annotation infrastructure

Low level access to FANC's CAVE annotation infrastructure

Usage

fanc_cave_client()

banc_cave_client()

Value

A reticulate R object wrapping the python CAVEclient.

Examples

# \donttest{
fcc=fanc_cave_client()
tables=fcc$annotation$get_tables()
fcc$materialize$get_table_metadata(tables[1])
#> $table_name
#> [1] "wing_motor_neuron_table_v0"
#> 
#> $schema
#> [1] "cell_type_local"
#> 
#> $created
#> [1] "2022-12-01T08:10:39.475152"
#> 
#> $valid
#> [1] TRUE
#> 
#> $id
#> [1] 31830
#> 
#> $aligned_volume
#> [1] "fanc_v4"
#> 
#> $schema_type
#> [1] "cell_type_local"
#> 
#> $user_id
#> [1] "690"
#> 
#> $description
#> [1] "left and right wing motor neurons, with their identities. Identites that are still under review are denoted with a \"_u\" suffix. The left wing MNs will be in Azevedo et. al., in prep and the right side MNs are paired across the midline by eye. Table is managed by Ellen Lesser, Tuthill Lab."
#> 
#> $notice_text
#> NULL
#> 
#> $reference_table
#> NULL
#> 
#> $flat_segmentation_source
#> NULL
#> 
#> $write_permission
#> [1] "PRIVATE"
#> 
#> $read_permission
#> [1] "PUBLIC"
#> 
#> $last_modified
#> [1] "2023-10-26T22:28:02.250171"
#> 
#> $segmentation_source
#> NULL
#> 
#> $pcg_table_name
#> [1] "mar2021_prod"
#> 
#> $last_updated
#> [1] "2024-08-03T21:00:00.269838"
#> 
#> $voxel_resolution
#> [1]  4.3  4.3 45.0
#> 
# }