Summarise the connectivity of FANC neurons

fanc_partners returns details of each unitary synaptic connection (including its xyz location).

Usage

fanc_partner_summary(
  rootids,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  datastack_name = NULL,
  ...
)

fanc_partners(rootids, partners = c("input", "output"), ...)

Arguments

rootids

Character vector specifying one or more flywire rootids. As a convenience this argument is passed to fanc_ids allowing you to pass in data.frames, flywire URLs or simple ids.

partners

Whether to fetch input or output synapses or both.

threshold

For flywire_partner_summary only return partners with greater than this number of connections to the query neuron(s) (default of 0 returns all connections)

remove_autapses

For flywire_partner_summary whether to remove autapses (defaults to TRUE)

cleft.threshold

A threshold for the cleft score calculated by Buhmann et al 2019 (default 0, we have used 30-100 to increase specificity)

datastack_name

An optional CAVE datastack_name. If unset a sensible default is chosen.

...

Additional arguments passed to pbsapply

Value

a data.frame

Details

note that the rootids you pass in must be up to date. See example.

See also

Examples

# NB id must be up to date
sample_id=fanc_latestid("648518346481082458")
head(fanc_partner_summary(sample_id))
#> # A tibble: 6 × 3
#>   post_id            weight     n
#>   <chr>               <int> <int>
#> 1 648518346490269367    136     1
#> 2 648518346493875704    103     1
#> 3 648518346501611089     85     1
#> 4 648518346488813390     72     1
#> 5 648518346471756187     70     1
#> 6 648518346521763321     70     1
head(fanc_partner_summary(sample_id, partners='inputs'))
#> # A tibble: 6 × 3
#>   pre_id             weight     n
#>   <chr>               <int> <int>
#> 1 648518346494014715    108     1
#> 2 648518346489556055     42     1
#> 3 648518346477266376     40     1
#> 4 648518346490339932     23     1
#> 5 648518346491851887     21     1
#> 6 648518346486859725     20     1
if (FALSE) { # \dontrun{
# get the latest id for an outdate
fanc_partner_summary(fanc_latestid("648518346473954669"))

## open fanc/flywire scene containing top partners
library(dplyr)
fanc_partner_summary(fanc_latestid("648518346494405175"), partners='inputs') %>%
  slice_max(weight, n = 20) %>%
  fanc_scene(open=TRUE)
} # }
if (FALSE) { # \dontrun{
# plot input and output synapses of a neuron
nclear3d()
fpi=fanc_partners(fanc_latestid("648518346481082458"), partners='in')
points3d(fanc_raw2nm(fpi$post_pt_position), col='cyan')
fpo=fanc_partners(fanc_latestid("648518346481082458"), partners='out')
points3d(fanc_raw2nm(fpo$pre_pt_position), col='red')
} # }