Mirror points or other 3D objects along the FANC midline

symmetric_fanc transforms neurons, surfaces and other point data onto a symmetrised version of the FANC template brain, optionally mirroring across the midline.

mirror_fanc(x, ...)

symmetric_fanc(x, mirror = FALSE, ...)

Arguments

x
3D vertices (or object containing them) in FANC space Could be neuron, neuronlist, hxsurf etc.

...
additional arguments passed to xform and friends

mirror
Whether to mirror across the midline when using symmetric_fanc

Value

Transformed points/object

Details

These registration functions depend on an installation of the CMTK registration toolkit. See cmtk.bindir for details.

Examples

# \donttest{
# TODO
# }

if (FALSE) {
FANC.surf.symm <- symmetric_fanc(FANC.surf)
# plot the two meshes: before and after
wire3d(FANC.surf, col='grey')
wire3d(FANC.surf.symm, col='red')
}