Read one or more FANC neuron meshes

Read one or more FANC neuron meshes

Usage

read_fanc_meshes(ids, savedir = NULL, format = c("ply", "obj"), ...)

Arguments

ids

One or more root ids

savedir

An optional location to save downloaded meshes. This acts as a simple but effective cache since flywire neurons change id whenever they are edited.

format

whether to save meshes in Wavefront obj or Stanford poly format. obj is the default but ply is a simpler and more compact format.

...

Additional arguments passed to fafbseg::read_cloudvolume_meshes

Value

A neuronlist containing one or more mesh3d objects. See nat::read.neurons for details.

See also

Examples

# \donttest{
read_fanc_meshes("648518346482929060")
#>   downloading meshes
#>   parsing downloaded meshes
#> 'neuronlist' containing 1 'mesh3d' object and 'data.frame' with 0 vars [12.5 MB]
# }