Transform FANC to MANC

transforms neurons, surfaces and other point data onto from FANC to MANC space.

Usage

transform_fanc2manc(x, inverse = F, ...)

Arguments

x

an object to transform (calibrated in nm)

inverse

boolean flag that says whether to swap the registration

...

additional arguments passed to (passed on to xform)

Examples

if (FALSE) {
library(nat)
library(malevnc)
FANC.in.manc <- transform_fanc2manc(FANC.surf)
# plot MANC and FANC mesh for comparison
wire3d(malevnc::MANC.surf, col='grey', add=T)
wire3d(FANC.in.manc/1e3, col='blue',add=T)
}