Fetch change log information for one or more neurons

Fetch change log information for one or more neurons

fanc_change_log(x, ...)

Arguments

x
One or more fanc ids in any format understandable by ngl_segments

...
Additional argument passed to pbsapply and then on to flywire_fetch (via the private fanc_fetch).

Value

A list containing information on the changes applied to a given body

n_splits,n_mergers The number of splits/mergers applied to the neuron

user_info A data.frame containing user ids or emails and the number of splits/mergers

operations_ids An integer vector of operation ids

past_ids Previous ids associated with this body

Examples

# \donttest{
fanc_change_log("648518346473954669")

#> $n_splits
#> [1] 5
#> 
#> $n_mergers
#> [1] 3
#> 
#> $user_info
#>                             user n_splits n_mergers
#> 1                            121        1         0
#> 2 jtmaniatesselvin@g.harvard.edu        3         2
#> 3                             16        1         1
#> 
#> $operations_ids
#> [1] 24951  1549   868  1548   867   866   865   864
#> 
#> $past_ids
#>  [1] "648518346480772096" "648518346473954669" "648518346509284227"
#>  [4] "648518346517118762" "648518346480558813" "648518346489510873"
#>  [7] "648518346492263865" "648518346490778913" "648518346520326710"
#> [10] "648518346498254576" "648518346492120771" "648518346492883740"
#> 

# }