Fetch change log information for one or more neurons

Fetch change log information for one or more neurons

Usage

fanc_change_log(x, tz = "UTC", filtered = TRUE, OmitFailures = TRUE, ...)

Arguments

x

One or more fanc ids in any format understandable by ngl_segments

tz

Time zone for edit timestamps. Defaults to "UTC" i.e. Universal Time, Coordinated. Set to "" for your current timezone. See as.POSIXct for more details.

filtered

Whether to filter out edits unlikely to relate to the current state of the neuron (default TRUE, see details).

OmitFailures

Whether to omit neurons for which there is an API timeout or error. The default value (TRUE) will skip over errors, while NA) will result in a hard stop on error. See nlapply for more details.

...

Additional arguments passed to flywire_fetch

Value

a data.frame See fabseg::flywire_change_log for details

Details

As of August 2021 this is a simple wrapper of fabseg::flywire_change_log. For now the old (and less convenient format) available from the zetta API can be obtained with the private fancr:::fanc_change_log_zetta function.

Examples

# \donttest{
fanc_change_log("648518346473954669")
#>   operation_id           timestamp user_id
#> 0          864 2021-03-23 17:43:02      69
#> 1          865 2021-03-23 17:46:30      69
#> 2          866 2021-03-23 17:50:40      69
#> 3          867 2021-03-23 17:51:18      69
#> 4          868 2021-03-23 17:52:21      69
#> 5         1548 2021-04-15 01:40:21      69
#> 6         1549 2021-04-15 01:41:03      69
#> 7        24951 2021-07-01 06:05:28    1211
#>                         before_root_ids     after_root_ids is_merge
#> 0                    648518346492296936 648518346520326710    FALSE
#> 1                    648518346520326710 648518346490778913    FALSE
#> 2 648518346485415597 648518346490778913 648518346492883740     TRUE
#> 3 648518346492883740 648518346506053173 648518346492120771     TRUE
#> 4                    648518346492120771 648518346517118762    FALSE
#> 5                    648518346504349342 648518346489510873    FALSE
#> 6 648518346489510873 648518346517118762 648518346498254576     TRUE
#> 7                    648518346498254576 648518346473954669    FALSE
#>       user_name  user_affiliation
#> 0 Jasper Phelps Wei-Chung Lee Lab
#> 1 Jasper Phelps Wei-Chung Lee Lab
#> 2 Jasper Phelps Wei-Chung Lee Lab
#> 3 Jasper Phelps Wei-Chung Lee Lab
#> 4 Jasper Phelps Wei-Chung Lee Lab
#> 5 Jasper Phelps Wei-Chung Lee Lab
#> 6 Jasper Phelps Wei-Chung Lee Lab
#> 7    Arti Yadav Greg Jefferis Lab
# }