Summarise the connectivity of FANC neurons

Summarise the connectivity of FANC neurons

fanc_partner_summary(
  rootid,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  datastack_name = NULL,
  ...
)

Arguments

rootid
Character vector specifying one or more flywire rootids. As a convenience for flywire_partner_summary this argument is passed to ngl_segments allowing you to pass in

partners
Whether to fetch input or output synapses or both.

threshold
For flywire_partner_summary only return partners with greater than this number of connections to the query neuron(s) (default of 0 returns all connections)

remove_autapses
For flywire_partner_summary whether to remove autapses (defaults to TRUE)

cleft.threshold
A threshold for the cleft score calculated by Buhmann et al 2019 (default 0, we have used 30-100 to increase specificity)

datastack_name
An optional CAVE datastack_name. If unset a sensible default is chosen.

...
Additional arguments passed to pbsapply

Value

a data.frame

Details

note that the rootids you pass in must be up to date. See example.

See also

Examples

head(fanc_partner_summary("648518346494405175"))
#> # A tibble: 6 × 3
#>   post_id            weight     n
#>   <chr>               <int> <int>
#> 1 648518346531154458     88     1
#> 2 648518346489890587     62     1
#> 3 648518346493661944     61     1
#> 4 648518346514203972     59     1
#> 5 648518346482131092     54     1
#> 6 648518346479673061     52     1
head(fanc_partner_summary("648518346494405175", partners='inputs'))
#> # A tibble: 6 × 3
#>   pre_id             weight     n
#>   <chr>               <int> <int>
#> 1 648518346518826577     72     1
#> 2 648518346489556055     27     1
#> 3 648518346492191976     25     1
#> 4 648518346517181733     24     1
#> 5 648518346475288756     21     1
#> 6 648518346486426344     21     1
if (FALSE) {
# get the latest id for an outdate
fanc_partner_summary(fanc_latestid("648518346473954669"))

## open fanc/flywire scene containing top partners
library(dplyr)
fanc_partner_summary("648518346494405175", partners='inputs') %>%
  slice_max(weight, n = 20) %>%
  fanc_scene(open=TRUE)
}