Read L2 skeleton or dotprops for FANC neurons using fafbseg-py

fanc_read_l2skel reads one or more neurons as simplified L2 skeletons.

fanc_read_l2dp reads one or more neurons as simplified dotprops format. See read_l2skel.

Usage

fanc_read_l2dp(id, OmitFailures = TRUE, dataset = NULL, ...)

fanc_read_l2skel(id, OmitFailures = TRUE, dataset = NULL, ...)

Arguments

id

One or more flywire ids

OmitFailures

Whether or not to drop neurons that cannot be read from the results (rather than erroring out). Default TRUE.

dataset

An optional CAVE dataset name (expert use only, by default will choose the standard FANC dataset). See details.

...

Additional arguments passed to the fafbseg.flywire.l2_skeleton or fafbseg.flywire.l2_dotpropsfunctions.

Value

a neuronlist containing one or more neuron or dotprops objects. Note that neurons will be calibrated in nm while dotprops will be calibrated in microns.

Details

fanc_read_l2dp uses a special data structure for rapid download of the dotprops version of neurons required for NBLASTing. It leverages the python navis / fafbseg-py packages and you will need to install these, typically using the simple_python function.

fanc_read_l2skel treats the dataset argument a little differently than fanc_read_l2dp because it actually needs to identify two data sources a CAVE data

See read_l2skel for additional details of

Examples

if (FALSE) { # \dontrun{
# one time install of necessary python packages
fafbseg::simple_python(pkgs="fafbseg")

dnp42=c("648518346507131167", "648518346485772414")
dnp42.latest=fanc_latestid(dnp42)
dnp42.dps <- fanc_read_l2dp(dnp42.latest)

# plot those
nclear3d()
plot3d(dnp42.dps, lwd=3)
# nb dotprops are always in microns
wire3d(FANC.surf/1e3, col='grey')

nclear3d()
dnp42.skel <- fanc_read_l2skel(dnp42.latest)
plot3d(dnp42.skel, lwd=2)
# nb neuron skeletons are in nm
wire3d(FANC.surf, col='grey')
} # }